requirements.txt
python-telegram-bot>=20.0
Pillow
moviepy
